requirements.txt
selenium
selenium-stealth
rich
requests
beautifulsoup4
python-dotenv
